requirements.txt
pyspark==3.4.4
requests
PyYAML
python-dotenv
psycopg2-binary
pytest
apache-airflow
pendulum